Move convert encoder out of IP addr module.
